<zero __repr__ fix>
In using the __repr__ function to create the string representation of the node, I had put in an 'or None' when assigning the data variables. 
This caused problems, because I thought when it was printing a node [data: None, next: 1]  That 0 was not being kept as a data and None was instead being placed in the node.  It was the data in the string representation that was None, and not the actual data.   I've included a stackoverflow link that explains in further detail.